🧹 Centralize duplicate formatBytes, ISOLATION_STATUS_COLORS, DEFAULT_PAGE_SIZE (#10353)
- Replace local formatBytes in NetworkTraceCard with import from lib/formatters
- Extract ISOLATION_STATUS_COLORS to multi-tenancy shared.ts, remove from 3 files
- Replace local DEFAULT_PAGE_SIZE = 5 in ActiveAlerts, IntotoSupplyChain, VClusterStatus with import from lib/constants/ui
- Consolidate IsolationStatus/IsolationLevel types into shared.ts, re-export from hooks
