+++ web/src/components/cards/gadget/NetworkTraceCard.tsx
@@ -2,6 +2,7 @@ import { Network, ArrowRight } from 'lucide-react'
 import { useCachedNetworkTraces } from '../../../hooks/useGadget'
 import { useCardLoadingState, useCardDemoState } from '../CardDataContext'
 import { ClusterBadge } from '../../ui/ClusterBadge'
+import { formatBytes } from '../../../lib/formatters'
 
 interface NetworkTraceCardProps {
   config?: Record<string, unknown>
@@ -76,8 +77,3 @@ export function NetworkTraceCard({ config }: NetworkTraceCardProps) {
   )
 }
 
-function formatBytes(bytes: number): string {
-  if (bytes < 1024) return `${bytes}B`
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)}KB`
-  return `${(bytes / (1024 * 1024)).toFixed(1)}MB`
-}
