+++ web/src/components/cards/multi-tenancy/tenant-isolation-setup/TenantIsolationSetup.tsx
@@ -23,7 +23,9 @@ import {
   KUBEVIRT_INSTALL_PROMPT } from './missionPrompts'
 import { loadMissionPrompt } from '../missionLoader'
 
-import type { ComponentReadiness, IsolationLevel, IsolationStatus } from './useTenantIsolationSetup'
+import type { ComponentReadiness } from './useTenantIsolationSetup'
+import type { IsolationLevel, IsolationStatus } from '../shared'
+import { ISOLATION_STATUS_COLORS } from '../shared'
 
 /** Icon map for component keys */
 const COMPONENT_ICONS: Record<string, React.ComponentType<{ className?: string }>> = {
@@ -50,12 +52,6 @@ function IsolationStatusIcon({ status }: { status: IsolationStatus }) {
       return <XCircle className="w-3.5 h-3.5 text-zinc-500" />
   }
 }
-
-/** Color classes for isolation status */
-const ISOLATION_STATUS_COLORS: Record<IsolationStatus, string> = {
-  ready: 'text-green-400',
-  degraded: 'text-orange-400',
-  missing: 'text-zinc-500' }
 
 /** Component readiness badge */
 function ReadinessBadge({ component, onInstall }: {
