+++ web/src/components/cards/multi-tenancy/tenant-isolation-setup/useTenantIsolationSetup.ts
@@ -19,13 +19,8 @@ export interface ComponentReadiness {
   health: string
 }
 
-export type IsolationStatus = 'ready' | 'missing' | 'degraded'
-
-export interface IsolationLevel {
-  type: string
-  status: IsolationStatus
-  provider: string
-}
+export type { IsolationStatus, IsolationLevel } from '../shared'
+import type { IsolationStatus, IsolationLevel } from '../shared'
 
 export interface TenantIsolationSetupData {
   components: ComponentReadiness[]
